test: normalize localPath slashes on Windows
