+++ lib/resolver.js
@@ -1,12 +1,12 @@
 import { join, sep as dirSep } from 'node:path';
 
 import { getIndex, getKind, getRealpath, isReadable } from './fs-utils.js';
+import { PathMatcher } from './path-matcher.js';
 import { fwdSlash, trimSlash } from './utils.js';
 
 /**
 @typedef {import('./types.js').DirIndexItem} DirIndexItem
 @typedef {import('./types.js').FSEntryBase} FSEntryBase
-@typedef {import('./types.js').FSEntryKind} FSEntryKind
 @typedef {import('./types.js').ResolveResult} ResolveResult
 @typedef {import('./types.js').ServerOptions} ServerOptions
 **/
@@ -83,7 +83,7 @@ export class FileResolver {
 
 			const enabled = resource.kind === 'file' || (resource.kind === 'dir' && this.#dirList);
 
-			if (enabled && this.allowedPath(localPath)) {
+			if (enabled && this.allowedLocalPath(localPath)) {
 				const readable = await isReadable(resource.filePath, resource.kind);
 				result.status = readable ? 200 : 403;
 			} else if (isSymlink) {
@@ -92,59 +92,6 @@ export class FileResolver {
 		}
 
 		return result;
-	}
-
-	/**
-	 * Locate alternative files that can be served for a resource,
-	 * using the config for extensions and index file lookup.
-	 * @param {string} targetPath
-	 * @returns {Promise<FSEntryBase>}
-	 */
-	async locateFile(targetPath) {
-		const targetKind = await getKind(targetPath);
-
-		if (targetKind === 'file' || targetKind === 'link') {
-			return { kind: targetKind, filePath: targetPath };
-		}
-
-		/** @type {string[]} */
-		let candidates = [];
-		if (targetKind === 'dir' && this.#dirFile.length) {
-			candidates = this.#dirFile.map((name) => join(targetPath, name));
-		} else if (targetKind === null && this.#ext.length) {
-			candidates = this.#ext.map((ext) => targetPath + ext);
-		}
-
-		for (const filePath of candidates) {
-			const kind = await getKind(filePath);
-			if (kind === 'file' || kind === 'link') {
-				return { kind, filePath };
-			}
-		}
-
-		return { kind: targetKind, filePath: targetPath };
-	}
-
-	/**
-	 * @param {string | null} localPath
-	 * @returns {boolean}
-	 */
-	allowedPath(localPath) {
-		if (typeof localPath === 'string') {
-			return this.#excludeMatcher.test(localPath) === false;
-		}
-		return false;
-	}
-
-	/**
-	 * @param {string} filePath
-	 * @returns {string | null}
-	 */
-	localPath(filePath) {
-		if (this.withinRoot(filePath)) {
-			return filePath.slice(this.#root.length + 1);
-		}
-		return null;
 	}
 
 	/**
@@ -159,7 +106,7 @@ export class FileResolver {
 
 		for (const { kind, filePath } of await getIndex(dirPath)) {
 			const localPath = this.localPath(filePath);
-			if (kind != null && this.allowedPath(localPath)) {
+			if (kind != null && this.allowedLocalPath(localPath)) {
 				items.push({ filePath, localPath, kind });
 			}
 		}
@@ -186,6 +133,48 @@ export class FileResolver {
 	}
 
 	/**
+	 * Locate alternative files that can be served for a resource,
+	 * using the config for extensions and index file lookup.
+	 * @param {string} fullPath
+	 * @returns {Promise<FSEntryBase>}
+	 */
+	async locateFile(fullPath) {
+		const targetKind = await getKind(fullPath);
+
+		if (targetKind === 'file' || targetKind === 'link') {
+			return { kind: targetKind, filePath: fullPath };
+		}
+
+		/** @type {string[]} */
+		let candidates = [];
+		if (targetKind === 'dir' && this.#dirFile.length) {
+			candidates = this.#dirFile.map((name) => join(fullPath, name));
+		} else if (targetKind === null && this.#ext.length) {
+			candidates = this.#ext.map((ext) => fullPath + ext);
+		}
+
+		for (const filePath of candidates) {
+			const kind = await getKind(filePath);
+			if (kind === 'file' || kind === 'link') {
+				return { kind, filePath };
+			}
+		}
+
+		return { kind: targetKind, filePath: fullPath };
+	}
+
+	/**
+	 * @param {string | null} localPath
+	 * @returns {boolean}
+	 */
+	allowedLocalPath(localPath) {
+		if (typeof localPath === 'string') {
+			return this.#excludeMatcher.test(localPath) === false;
+		}
+		return false;
+	}
+
+	/**
 	 * @param {string} url
 	 * @returns {string | null}
 	 */
@@ -196,6 +185,17 @@ export class FileResolver {
 				return path.startsWith('/') ? path : `/${path}`;
 			}
 		} catch {}
+		return null;
+	}
+
+	/**
+	 * @param {string} fullPath
+	 * @returns {string | null}
+	 */
+	localPath(fullPath) {
+		if (this.withinRoot(fullPath)) {
+			return fullPath.slice(this.#root.length + 1);
+		}
 		return null;
 	}
 
@@ -224,113 +224,12 @@ export class FileResolver {
 	}
 
 	/**
-	 * @param {string} filePath
+	 * @param {string} fullPath
 	 * @returns {boolean}
 	 */
-	withinRoot(filePath) {
-		if (filePath.includes('..')) return false;
+	withinRoot(fullPath) {
+		if (fullPath.includes('..')) return false;
 		const prefix = this.#root + dirSep;
-		return filePath === this.#root || filePath.startsWith(prefix);
-	}
-}
-
-/**
- * @typedef {{ caseSensitive: boolean }} PathMatcherOptions
- */
-export class PathMatcher {
-	/** @type {Array<string | RegExp>} */
-	#positive = [];
-
-	/** @type {Array<string | RegExp>} */
-	#negative = [];
-
-	/** @type {PathMatcherOptions} */
-	#options = {
-		caseSensitive: true,
-	};
-
-	/**
-	 * @param {string[]} patterns
-	 * @param {Partial<PathMatcherOptions>} [options]
-	 */
-	constructor(patterns, { caseSensitive } = {}) {
-		if (typeof caseSensitive === 'boolean') {
-			this.#options.caseSensitive = caseSensitive;
-		}
-		for (const input of patterns) {
-			if (typeof input !== 'string') continue;
-			const isNegative = input.startsWith('!');
-			const trimmedInput = input.slice(isNegative ? 1 : 0);
-			const pattern = trimmedInput.length > 0 ? this.#parse(trimmedInput) : null;
-			if (pattern != null) {
-				(isNegative ? this.#negative : this.#positive).push(pattern);
-			}
-		}
-	}
-
-	/**
-	 * @param {string} filePath
-	 * @returns {boolean}
-	 */
-	test(filePath) {
-		if (this.#positive.length === 0) {
-			return false;
-		}
-		const segments = fwdSlash(filePath).split('/').filter(Boolean);
-		const matched = this.#matchSegments(segments);
-		return matched.length > 0;
-	}
-
-	get rules() {
-		return structuredClone({ positive: this.#positive, negative: this.#negative });
-	}
-
-	/**
-	 * @param {string} input
-	 * @returns {string | RegExp | null}
-	 */
-	#parse(input) {
-		if (this.#options.caseSensitive === false) {
-			input = input.toLowerCase();
-		}
-		if (input.includes('/') || input.includes('\\')) {
-			return null;
-		} else if (input.includes('*')) {
-			const toEscape = /([\[\]\(\)\|\^\$\.\+\?])/g;
-			const re = input.replace(toEscape, '\\$1').replace(/\*/g, '[^/]*');
-			return new RegExp(re);
-		}
-		return input;
-	}
-
-	/**
-	 * @param {string | RegExp} pattern
-	 * @param {string} value
-	 * @returns {boolean}
-	 */
-	#matchPattern(pattern, value) {
-		if (this.#options.caseSensitive === false) {
-			value = value.toLowerCase();
-		}
-		if (typeof pattern === 'string') {
-			return pattern === value;
-		} else if (pattern.test(value)) {
-			const matches = value.match(pattern);
-			return matches != null && matches[0] === value;
-		}
-		return false;
-	}
-
-	/**
-	 * @param {string[]} segments
-	 * @returns {string[]}
-	 */
-	#matchSegments(segments) {
-		return segments.filter((segment) => {
-			const positive = this.#positive.some((pattern) => this.#matchPattern(pattern, segment));
-			if (!positive) return false;
-			const negative = this.#negative.some((pattern) => this.#matchPattern(pattern, segment));
-			return positive && !negative;
-		});
+		return fullPath === this.#root || fullPath.startsWith(prefix);
 	}
 }
