+++ lib/server.js
@@ -7,7 +7,8 @@ import { createGzip, gzipSync } from 'node:zlib';
 import { MAX_COMPRESS_SIZE, SUPPORTED_METHODS } from './constants.js';
 import { getContentType, typeForFilePath } from './content-type.js';
 import { dirListPage, errorPage } from './pages.js';
-import { FileResolver, PathMatcher } from './resolver.js';
+import { PathMatcher } from './path-matcher.js';
+import { FileResolver } from './resolver.js';
 import { headerCase } from './utils.js';
 
 /**
@@ -16,11 +17,8 @@ import { headerCase } from './utils.js';
 @typedef {import('node:http').Server} Server
 @typedef {import('node:http').ServerResponse} ServerResponse
 @typedef {import('./content-type.js').TypeResult} TypeResult
-@typedef {import('./types.js').DirIndexItem} DirIndexItem
-@typedef {import('./types.js').FSEntryKind} FSEntryKind
 @typedef {import('./types.js').ReqResMeta} ReqResMeta
 @typedef {import('./types.js').ResolvedFile} ResolvedFile
-@typedef {import('./types.js').ResolveResult} ResolveResult
 @typedef {import('./types.js').ServerOptions} ServerOptions
 @typedef {{
 	body?: string | Buffer | import('node:fs').ReadStream;
